Decouple snapshot code from restate-core
- Remove unused snapshot code from PPM handle
- Moves snapshot helper types into partition_store::snapshots
- Removes a couple of unnused dependencies from restate-core.
