+++ crates/partition-store/src/snapshots.rs
@@ -16,10 +16,42 @@ use serde::{Deserialize, Serialize};
 use serde_with::hex::Hex;
 use serde_with::{DeserializeAs, SerializeAs, serde_as};
 
-use restate_core::worker_api::SnapshotCreated;
 use restate_types::identifiers::{PartitionId, PartitionKey, SnapshotId};
 use restate_types::logs::{LogId, Lsn};
 use restate_types::nodes_config::ClusterFingerprint;
+
+#[derive(Debug, Clone, derive_more::Display)]
+#[display("{}", snapshot_id)]
+pub struct SnapshotCreated {
+    pub snapshot_id: SnapshotId,
+    pub log_id: LogId,
+    pub min_applied_lsn: Lsn,
+}
+
+#[derive(Debug, derive_more::Display)]
+#[display("{kind} for partition {partition_id}")]
+pub struct SnapshotError {
+    pub partition_id: PartitionId,
+    pub kind: SnapshotErrorKind,
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum SnapshotErrorKind {
+    #[error("Partition not found")]
+    PartitionNotFound,
+    #[error("Snapshot export in progress")]
+    SnapshotInProgress,
+    #[error("Partition Processor state does not permit snapshotting")]
+    InvalidState,
+    #[error("Snapshot repository is not configured")]
+    RepositoryNotConfigured,
+    #[error("Snapshot export failed for partition")]
+    Export(#[source] anyhow::Error),
+    #[error("Snapshot repository IO error")]
+    RepositoryIo(#[source] anyhow::Error),
+    #[error("Internal error")]
+    Internal(anyhow::Error),
+}
 
 #[derive(Debug, Clone, Copy, Default, Eq, PartialEq, Serialize, Deserialize)]
 pub enum SnapshotFormatVersion {
